# process: Deduplicate paths in parseMappings()