Reduce number of allocations significantly in `parseMappings()`. It also has some positive effects on CPU usage.

```
$ benchstat old.txt new.txt
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/ebpf-profiler/process
cpu: 12th Gen Intel(R) Core(TM) i7-12800H
                 │   old.txt   │              new.txt               │
                 │   sec/op    │   sec/op     vs base               │
ParseMappings-20   90.92µ ± 5%   85.19µ ± 2%  -6.31% (p=0.000 n=10)

                 │   old.txt    │               new.txt                │
                 │     B/op     │     B/op      vs base                │
ParseMappings-20   97.82Ki ± 0%   80.92Ki ± 0%  -17.27% (p=0.000 n=10)

                 │   old.txt   │              new.txt               │
                 │  allocs/op  │ allocs/op   vs base                │
ParseMappings-20   438.00 ± 0%   95.00 ± 0%  -78.31% (p=0.000 n=10)
```

The mappings are from a locally running Firefox. The benchmark code is:
```
func BenchmarkParseMappings(b *testing.B) {
	for i := 0; i < b.N; i++ {
		mappings, err := parseMappings(strings.NewReader(firefoxMappings))
		if err != nil {
			b.Fatal(err)
		}
		if len(mappings) == 0 {
			b.Fatal("expected non-empty mappings")
		}
	}
}
```

Just for reproducibility:
<details close>
<summary>Firefox mappings</summary>

```Go
var firefoxMappings = `55b90916b000-55b909192000 r--p 00000000 fd:01 63180418                   /usr/lib/firefox/firefox
55b909192000-55b90920c000 r-xp 00026000 fd:01 63180418                   /usr/lib/firefox/firefox
55b90920c000-55b90920f000 r--p 0009f000 fd:01 63180418                   /usr/lib/firefox/firefox
55b90920f000-55b909210000 rw-p 000a1000 fd:01 63180418                   /usr/lib/firefox/firefox
55b909210000-55b909212000 rw-p 00000000 00:00 0 
7f61fae00000-7f61fae02000 rw-p 00000000 00:00 0 
7f61fae02000-7f61fae03000 ---p 00000000 00:00 0 
7f61fae03000-7f61faeff000 rw-p 00000000 00:00 0 
7f61faeff000-7f61faf00000 ---p 00000000 00:00 0 
7f61faf4b000-7f61faf4c000 ---p 00000000 00:00 0 
7f61faf4c000-7f61faf8c000 rw-p 00000000 00:00 0 
7f61faf8c000-7f61faf8d000 ---p 00000000 00:00 0 
7f61faf8d000-7f61fafcd000 rw-p 00000000 00:00 0 
7f61fafcd000-7f61fafdb000 r--p 00000000 fd:01 63180427                   /usr/lib/firefox/libipcclientcerts.so
7f61fafdb000-7f61fb02a000 r-xp 0000d000 fd:01 63180427                   /usr/lib/firefox/libipcclientcerts.so
7f61fb02a000-7f61fb02d000 r--p 0005b000 fd:01 63180427                   /usr/lib/firefox/libipcclientcerts.so
7f61fb02d000-7f61fb02e000 rw-p 0005d000 fd:01 63180427                   /usr/lib/firefox/libipcclientcerts.so
7f61fb02e000-7f61fb032000 r--p 00000000 fd:01 56766108                   /usr/lib/x86_64-linux-gnu/libfreeblpriv3.so
7f61fb032000-7f61fb0eb000 r-xp 00004000 fd:01 56766108                   /usr/lib/x86_64-linux-gnu/libfreeblpriv3.so
7f61fb0eb000-7f61fb11a000 r--p 000bd000 fd:01 56766108                   /usr/lib/x86_64-linux-gnu/libfreeblpriv3.so
7f61fb11a000-7f61fb11c000 r--p 000ec000 fd:01 56766108                   /usr/lib/x86_64-linux-gnu/libfreeblpriv3.so
7f61fb11c000-7f61fb11d000 rw-p 000ee000 fd:01 56766108                   /usr/lib/x86_64-linux-gnu/libfreeblpriv3.so
7f61fb11d000-7f61fb121000 rw-p 00000000 00:00 0 
7f61fb121000-7f61fb148000 r--p 00000000 fd:01 56763797                   /usr/lib/x86_64-linux-gnu/libsqlite3.so.0.8.6
7f61fb148000-7f61fb258000 r-xp 00027000 fd:01 56763797                   /usr/lib/x86_64-linux-gnu/libsqlite3.so.0.8.6
7f61fb258000-7f61fb296000 r--p 00137000 fd:01 56763797                   /usr/lib/x86_64-linux-gnu/libsqlite3.so.0.8.6
7f61fb296000-7f61fb29c000 r--p 00175000 fd:01 56763797                   /usr/lib/x86_64-linux-gnu/libsqlite3.so.0.8.6
7f61fb29c000-7f61fb2a1000 rw-p 0017b000 fd:01 56763797                   /usr/lib/x86_64-linux-gnu/libsqlite3.so.0.8.6
7f61fb2a1000-7f61fb2a7000 r--p 00000000 fd:01 56766363                   /usr/lib/x86_64-linux-gnu/libsoftokn3.so
7f61fb2a7000-7f61fb2e4000 r-xp 00006000 fd:01 56766363                   /usr/lib/x86_64-linux-gnu/libsoftokn3.so
7f61fb2e4000-7f61fb2fc000 r--p 00043000 fd:01 56766363                   /usr/lib/x86_64-linux-gnu/libsoftokn3.so
7f61fb2fc000-7f61fb2fd000 r--p 0005b000 fd:01 56766363                   /usr/lib/x86_64-linux-gnu/libsoftokn3.so
7f61fb2fd000-7f61fb2fe000 rw-p 0005c000 fd:01 56766363                   /usr/lib/x86_64-linux-gnu/libsoftokn3.so
7f61fb2fe000-7f61fb302000 rw-p 00000000 00:00 0 
7f61fb302000-7f61fb303000 ---p 00000000 00:00 0 
7f61fb303000-7f61fb3ff000 rw-p 00000000 00:00 0 
7f61fb3ff000-7f61fb400000 ---p 00000000 00:00 0 
7f61fb400000-7f61fb407000 r--p 00000000 fd:01 56756176                   /usr/lib/x86_64-linux-gnu/libvpx.so.9.1.0
7f61fb407000-7f61fb718000 r-xp 00007000 fd:01 56756176                   /usr/lib/x86_64-linux-gnu/libvpx.so.9.1.0
7f61fb718000-7f61fb769000 r--p 00318000 fd:01 56756176                   /usr/lib/x86_64-linux-gnu/libvpx.so.9.1.0
7f61fb769000-7f61fb76c000 r--p 00369000 fd:01 56756176                   /usr/lib/x86_64-linux-gnu/libvpx.so.9.1.0
7f61fb76c000-7f61fb76d000 rw-p 0036c000 fd:01 56756176                   /usr/lib/x86_64-linux-gnu/libvpx.so.9.1.0
7f61fb76d000-7f61fb774000 rw-p 00000000 00:00 0 
7f61fb78f000-7f61fb790000 ---p 00000000 00:00 0 
7f61fb790000-7f61fb7d0000 rw-p 00000000 00:00 0 
7f61fb7d0000-7f61fb7d1000 ---p 00000000 00:00 0 
7f61fb7d1000-7f61fb811000 rw-p 00000000 00:00 0 
7f61fb811000-7f61fb853000 r--s 00000000 00:01 16399                      /memfd:mozilla-ipc (deleted)
7f61fb853000-7f61fb854000 ---p 00000000 00:00 0 
7f61fb854000-7f61fb894000 rw-p 00000000 00:00 0 
7f61fb894000-7f61fb8a2000 r--p 00000000 fd:01 56764741                   /usr/lib/x86_64-linux-gnu/libevent-2.1.so.7.0.1
7f61fb8a2000-7f61fb8d4000 r-xp 0000e000 fd:01 56764741                   /usr/lib/x86_64-linux-gnu/libevent-2.1.so.7.0.1
7f61fb8d4000-7f61fb8e7000 r--p 00040000 fd:01 56764741                   /usr/lib/x86_64-linux-gnu/libevent-2.1.so.7.0.1
7f61fb8e7000-7f61fb8e9000 r--p 00053000 fd:01 56764741                   /usr/lib/x86_64-linux-gnu/libevent-2.1.so.7.0.1
7f61fb8e9000-7f61fb8ea000 rw-p 00055000 fd:01 56764741                   /usr/lib/x86_64-linux-gnu/libevent-2.1.so.7.0.1
7f61fb8ea000-7f61fb8f8000 r--p 00000000 fd:01 56766364                   /usr/lib/x86_64-linux-gnu/libssl3.so
7f61fb8f8000-7f61fb93a000 r-xp 0000e000 fd:01 56766364                   /usr/lib/x86_64-linux-gnu/libssl3.so
7f61fb93a000-7f61fb94f000 r--p 00050000 fd:01 56766364                   /usr/lib/x86_64-linux-gnu/libssl3.so
7f61fb94f000-7f61fb953000 r--p 00065000 fd:01 56766364                   /usr/lib/x86_64-linux-gnu/libssl3.so
7f61fb953000-7f61fb954000 rw-p 00069000 fd:01 56766364                   /usr/lib/x86_64-linux-gnu/libssl3.so
7f61fb954000-7f61fb955000 rw-p 00000000 00:00 0 
7f61fb955000-7f61fb95f000 r--p 00000000 fd:01 56766361                   /usr/lib/x86_64-linux-gnu/libsmime3.so
7f61fb95f000-7f61fb97c000 r-xp 0000a000 fd:01 56766361                   /usr/lib/x86_64-linux-gnu/libsmime3.so
7f61fb97c000-7f61fb984000 r--p 00027000 fd:01 56766361                   /usr/lib/x86_64-linux-gnu/libsmime3.so
7f61fb984000-7f61fb988000 r--p 0002f000 fd:01 56766361                   /usr/lib/x86_64-linux-gnu/libsmime3.so
7f61fb988000-7f61fb989000 rw-p 00033000 fd:01 56766361                   /usr/lib/x86_64-linux-gnu/libsmime3.so
7f61fb989000-7f61fb9a5000 r--p 00000000 fd:01 56766109                   /usr/lib/x86_64-linux-gnu/libnss3.so
7f61fb9a5000-7f61fbaa0000 r-xp 0001c000 fd:01 56766109                   /usr/lib/x86_64-linux-gnu/libnss3.so
7f61fbaa0000-7f61fbae1000 r--p 00117000 fd:01 56766109                   /usr/lib/x86_64-linux-gnu/libnss3.so
7f61fbae1000-7f61fbae8000 r--p 00158000 fd:01 56766109                   /usr/lib/x86_64-linux-gnu/libnss3.so
7f61fbae8000-7f61fbae9000 rw-p 0015f000 fd:01 56766109                   /usr/lib/x86_64-linux-gnu/libnss3.so
7f61fbae9000-7f61fbaeb000 rw-p 00000000 00:00 0 
7f61fbaeb000-7f61fbb1b000 r--p 00000000 fd:01 56755565                   /usr/lib/x86_64-linux-gnu/libasound.so.2.0.0
7f61fbb1b000-7f61fbbba000 r-xp 00030000 fd:01 56755565                   /usr/lib/x86_64-linux-gnu/libasound.so.2.0.0
7f61fbbba000-7f61fbbf7000 r--p 000cf000 fd:01 56755565                   /usr/lib/x86_64-linux-gnu/libasound.so.2.0.0
7f61fbbf7000-7f61fbbff000 r--p 0010c000 fd:01 56755565                   /usr/lib/x86_64-linux-gnu/libasound.so.2.0.0
7f61fbbff000-7f61fbc00000 rw-p 00114000 fd:01 56755565                   /usr/lib/x86_64-linux-gnu/libasound.so.2.0.0
7f61fbc00000-7f61fe7b3000 r--p 00000000 fd:01 63180439                   /usr/lib/firefox/libxul.so
7f61fe7b3000-7f6204c3b000 r-xp 02bb2000 fd:01 63180439                   /usr/lib/firefox/libxul.so
7f6204c3b000-7f6205112000 r--p 09039000 fd:01 63180439                   /usr/lib/firefox/libxul.so
7f6205112000-7f62051aa000 rw-p 0950f000 fd:01 63180439                   /usr/lib/firefox/libxul.so
7f62051aa000-7f6205257000 rw-p 00000000 00:00 0 
7f620527c000-7f6205289000 r--p 00000000 fd:01 56766360                   /usr/lib/x86_64-linux-gnu/libnssutil3.so
7f6205289000-7f620529b000 r-xp 0000d000 fd:01 56766360                   /usr/lib/x86_64-linux-gnu/libnssutil3.so
7f620529b000-7f62052a8000 r--p 0001f000 fd:01 56766360                   /usr/lib/x86_64-linux-gnu/libnssutil3.so
7f62052a8000-7f62052af000 r--p 0002c000 fd:01 56766360                   /usr/lib/x86_64-linux-gnu/libnssutil3.so
7f62052af000-7f62052b0000 rw-p 00033000 fd:01 56766360                   /usr/lib/x86_64-linux-gnu/libnssutil3.so
7f62052c6000-7f62052d2000 rw-p 00000000 00:00 0 
7f62052d2000-7f62052d5000 r--p 00000000 fd:01 56755779                   /usr/lib/x86_64-linux-gnu/libcap.so.2.66
7f62052d5000-7f62052da000 r-xp 00003000 fd:01 56755779                   /usr/lib/x86_64-linux-gnu/libcap.so.2.66
7f62052da000-7f62052dc000 r--p 00008000 fd:01 56755779                   /usr/lib/x86_64-linux-gnu/libcap.so.2.66
7f62052dc000-7f62052dd000 r--p 0000a000 fd:01 56755779                   /usr/lib/x86_64-linux-gnu/libcap.so.2.66
7f62052dd000-7f62052de000 rw-p 0000b000 fd:01 56755779                   /usr/lib/x86_64-linux-gnu/libcap.so.2.66
7f62052de000-7f62052df000 r--p 00000000 fd:01 56758033                   /usr/lib/x86_64-linux-gnu/libbrotlicommon.so.1.1.0
7f62052df000-7f62052e0000 r-xp 00001000 fd:01 56758033                   /usr/lib/x86_64-linux-gnu/libbrotlicommon.so.1.1.0
7f62052e0000-7f62052ff000 r--p 00002000 fd:01 56758033                   /usr/lib/x86_64-linux-gnu/libbrotlicommon.so.1.1.0
7f62052ff000-7f6205300000 r--p 00021000 fd:01 56758033                   /usr/lib/x86_64-linux-gnu/libbrotlicommon.so.1.1.0
7f6205300000-7f6205301000 rw-p 00022000 fd:01 56758033                   /usr/lib/x86_64-linux-gnu/libbrotlicommon.so.1.1.0
7f6205301000-7f620531d000 r--p 00000000 fd:01 56763774                   /usr/lib/x86_64-linux-gnu/libsystemd.so.0.40.0
7f620531d000-7f62053cc000 r-xp 0001c000 fd:01 56763774                   /usr/lib/x86_64-linux-gnu/libsystemd.so.0.40.0
7f62053cc000-7f6205407000 r--p 000cb000 fd:01 56763774                   /usr/lib/x86_64-linux-gnu/libsystemd.so.0.40.0
7f6205407000-7f6205414000 r--p 00105000 fd:01 56763774                   /usr/lib/x86_64-linux-gnu/libsystemd.so.0.40.0
7f6205414000-7f6205415000 rw-p 00112000 fd:01 56763774                   /usr/lib/x86_64-linux-gnu/libsystemd.so.0.40.0
7f6205415000-7f6205416000 rw-p 00000000 00:00 0 
7f6205416000-7f6205418000 r--p 00000000 fd:01 56758505                   /usr/lib/x86_64-linux-gnu/libXdmcp.so.6.0.0
7f6205418000-7f620541a000 r-xp 00002000 fd:01 56758505                   /usr/lib/x86_64-linux-gnu/libXdmcp.so.6.0.0
7f620541a000-7f620541c000 r--p 00004000 fd:01 56758505                   /usr/lib/x86_64-linux-gnu/libXdmcp.so.6.0.0
7f620541c000-7f620541d000 r--p 00005000 fd:01 56758505                   /usr/lib/x86_64-linux-gnu/libXdmcp.so.6.0.0
7f620541d000-7f620541e000 rw-p 00006000 fd:01 56758505                   /usr/lib/x86_64-linux-gnu/libXdmcp.so.6.0.0
7f620541e000-7f620541f000 r--p 00000000 fd:01 56758412                   /usr/lib/x86_64-linux-gnu/libXau.so.6.0.0
7f620541f000-7f6205420000 r-xp 00001000 fd:01 56758412                   /usr/lib/x86_64-linux-gnu/libXau.so.6.0.0
7f6205420000-7f6205421000 r--p 00002000 fd:01 56758412                   /usr/lib/x86_64-linux-gnu/libXau.so.6.0.0
7f6205421000-7f6205422000 r--p 00002000 fd:01 56758412                   /usr/lib/x86_64-linux-gnu/libXau.so.6.0.0
7f6205422000-7f6205423000 rw-p 00003000 fd:01 56758412                   /usr/lib/x86_64-linux-gnu/libXau.so.6.0.0
7f6205423000-7f6205424000 r--p 00000000 fd:01 56764113                   /usr/lib/x86_64-linux-gnu/libbrotlidec.so.1.1.0
7f6205424000-7f620542c000 r-xp 00001000 fd:01 56764113                   /usr/lib/x86_64-linux-gnu/libbrotlidec.so.1.1.0
7f620542c000-7f620542f000 r--p 00009000 fd:01 56764113                   /usr/lib/x86_64-linux-gnu/libbrotlidec.so.1.1.0
7f620542f000-7f6205430000 r--p 0000b000 fd:01 56764113                   /usr/lib/x86_64-linux-gnu/libbrotlidec.so.1.1.0
7f6205430000-7f6205431000 rw-p 0000c000 fd:01 56764113                   /usr/lib/x86_64-linux-gnu/libbrotlidec.so.1.1.0
7f6205431000-7f6205433000 r--p 00000000 fd:01 56754219                   /usr/lib/x86_64-linux-gnu/libbz2.so.1.0.4
7f6205433000-7f6205440000 r-xp 00002000 fd:01 56754219                   /usr/lib/x86_64-linux-gnu/libbz2.so.1.0.4
7f6205440000-7f6205442000 r--p 0000f000 fd:01 56754219                   /usr/lib/x86_64-linux-gnu/libbz2.so.1.0.4
7f6205442000-7f6205443000 r--p 00010000 fd:01 56754219                   /usr/lib/x86_64-linux-gnu/libbz2.so.1.0.4
7f6205443000-7f6205444000 rw-p 00011000 fd:01 56754219                   /usr/lib/x86_64-linux-gnu/libbz2.so.1.0.4
7f6205444000-7f6205446000 r--p 00000000 fd:01 56786448                   /usr/lib/x86_64-linux-gnu/libdatrie.so.1.4.0
7f6205446000-7f620544a000 r-xp 00002000 fd:01 56786448                   /usr/lib/x86_64-linux-gnu/libdatrie.so.1.4.0
7f620544a000-7f620544c000 r--p 00006000 fd:01 56786448                   /usr/lib/x86_64-linux-gnu/libdatrie.so.1.4.0
7f620544c000-7f620544d000 r--p 00007000 fd:01 56786448                   /usr/lib/x86_64-linux-gnu/libdatrie.so.1.4.0
7f620544d000-7f620544e000 rw-p 00008000 fd:01 56786448                   /usr/lib/x86_64-linux-gnu/libdatrie.so.1.4.0
7f620544e000-7f6205458000 r--p 00000000 fd:01 56755667                   /usr/lib/x86_64-linux-gnu/libblkid.so.1.1.0
7f6205458000-7f6205494000 r-xp 0000a000 fd:01 56755667                   /usr/lib/x86_64-linux-gnu/libblkid.so.1.1.0
7f6205494000-7f62054a7000 r--p 00046000 fd:01 56755667                   /usr/lib/x86_64-linux-gnu/libblkid.so.1.1.0
7f62054a7000-7f62054ad000 r--p 00059000 fd:01 56755667                   /usr/lib/x86_64-linux-gnu/libblkid.so.1.1.0
7f62054ad000-7f62054ae000 rw-p 0005f000 fd:01 56755667                   /usr/lib/x86_64-linux-gnu/libblkid.so.1.1.0
7f62054ae000-7f62054bc000 r--p 00000000 fd:01 56755806                   /usr/lib/x86_64-linux-gnu/libdbus-1.so.3.38.3
7f62054bc000-7f62054ed000 r-xp 0000e000 fd:01 56755806                   /usr/lib/x86_64-linux-gnu/libdbus-1.so.3.38.3
7f62054ed000-7f6205502000 r--p 0003f000 fd:01 56755806                   /usr/lib/x86_64-linux-gnu/libdbus-1.so.3.38.3
7f6205502000-7f6205504000 r--p 00053000 fd:01 56755806                   /usr/lib/x86_64-linux-gnu/libdbus-1.so.3.38.3
7f6205504000-7f6205505000 rw-p 00055000 fd:01 56755806                   /usr/lib/x86_64-linux-gnu/libdbus-1.so.3.38.3
7f6205505000-7f6205516000 r--p 00000000 fd:01 56763174                   /usr/lib/x86_64-linux-gnu/libatspi.so.0.0.1
7f6205516000-7f620552e000 r-xp 00011000 fd:01 56763174                   /usr/lib/x86_64-linux-gnu/libatspi.so.0.0.1
7f620552e000-7f620553c000 r--p 00029000 fd:01 56763174                   /usr/lib/x86_64-linux-gnu/libatspi.so.0.0.1
7f620553c000-7f6205540000 r--p 00036000 fd:01 56763174                   /usr/lib/x86_64-linux-gnu/libatspi.so.0.0.1
7f6205540000-7f6205541000 rw-p 0003a000 fd:01 56763174                   /usr/lib/x86_64-linux-gnu/libatspi.so.0.0.1
7f6205541000-7f6205545000 r--p 00000000 fd:01 56757433                   /usr/lib/x86_64-linux-gnu/libjpeg.so.62.3.0
7f6205545000-7f6205594000 r-xp 00004000 fd:01 56757433                   /usr/lib/x86_64-linux-gnu/libjpeg.so.62.3.0
7f6205594000-7f62055d6000 r--p 00053000 fd:01 56757433                   /usr/lib/x86_64-linux-gnu/libjpeg.so.62.3.0
7f62055d6000-7f62055d7000 r--p 00095000 fd:01 56757433                   /usr/lib/x86_64-linux-gnu/libjpeg.so.62.3.0
7f62055d7000-7f62055d8000 rw-p 00096000 fd:01 56757433                   /usr/lib/x86_64-linux-gnu/libjpeg.so.62.3.0
7f62055d8000-7f62055dc000 r--p 00000000 fd:01 56754304                   /usr/lib/x86_64-linux-gnu/libexpat.so.1.10.0
7f62055dc000-7f62055f8000 r-xp 00004000 fd:01 56754304                   /usr/lib/x86_64-linux-gnu/libexpat.so.1.10.0
7f62055f8000-7f6205600000 r--p 00020000 fd:01 56754304                   /usr/lib/x86_64-linux-gnu/libexpat.so.1.10.0
7f6205600000-7f6205602000 r--p 00028000 fd:01 56754304                   /usr/lib/x86_64-linux-gnu/libexpat.so.1.10.0
7f6205602000-7f6205603000 rw-p 0002a000 fd:01 56754304                   /usr/lib/x86_64-linux-gnu/libexpat.so.1.10.0
7f6205603000-7f620560e000 r--p 00000000 fd:01 56759633                   /usr/lib/x86_64-linux-gnu/libpixman-1.so.0.44.0
7f620560e000-7f6205695000 r-xp 0000b000 fd:01 56759633                   /usr/lib/x86_64-linux-gnu/libpixman-1.so.0.44.0
7f6205695000-7f62056a7000 r--p 00092000 fd:01 56759633                   /usr/lib/x86_64-linux-gnu/libpixman-1.so.0.44.0
7f62056a7000-7f62056af000 r--p 000a4000 fd:01 56759633                   /usr/lib/x86_64-linux-gnu/libpixman-1.so.0.44.0
7f62056af000-7f62056b0000 rw-p 000ac000 fd:01 56759633                   /usr/lib/x86_64-linux-gnu/libpixman-1.so.0.44.0
7f62056b0000-7f62056b1000 r--p 00000000 fd:01 56755461                   /usr/lib/x86_64-linux-gnu/libxcb-shm.so.0.0.0
7f62056b1000-7f62056b2000 r-xp 00001000 fd:01 56755461                   /usr/lib/x86_64-linux-gnu/libxcb-shm.so.0.0.0
7f62056b2000-7f62056b3000 r--p 00002000 fd:01 56755461                   /usr/lib/x86_64-linux-gnu/libxcb-shm.so.0.0.0
7f62056b3000-7f62056b4000 r--p 00002000 fd:01 56755461                   /usr/lib/x86_64-linux-gnu/libxcb-shm.so.0.0.0
7f62056b4000-7f62056b5000 rw-p 00003000 fd:01 56755461                   /usr/lib/x86_64-linux-gnu/libxcb-shm.so.0.0.0
7f62056b5000-7f62056ba000 r--p 00000000 fd:01 56760888                   /usr/lib/x86_64-linux-gnu/libxcb-render.so.0.0.0
7f62056ba000-7f62056c0000 r-xp 00005000 fd:01 56760888                   /usr/lib/x86_64-linux-gnu/libxcb-render.so.0.0.0
7f62056c0000-7f62056c2000 r--p 0000b000 fd:01 56760888                   /usr/lib/x86_64-linux-gnu/libxcb-render.so.0.0.0
7f62056c2000-7f62056c3000 r--p 0000d000 fd:01 56760888                   /usr/lib/x86_64-linux-gnu/libxcb-render.so.0.0.0
7f62056c3000-7f62056c4000 rw-p 0000e000 fd:01 56760888                   /usr/lib/x86_64-linux-gnu/libxcb-render.so.0.0.0
7f62056c4000-7f62056d0000 r--p 00000000 fd:01 56754381                   /usr/lib/x86_64-linux-gnu/libxcb.so.1.1.0
7f62056d0000-7f62056e4000 r-xp 0000c000 fd:01 56754381                   /usr/lib/x86_64-linux-gnu/libxcb.so.1.1.0
7f62056e4000-7f62056ed000 r--p 00020000 fd:01 56754381                   /usr/lib/x86_64-linux-gnu/libxcb.so.1.1.0
7f62056ed000-7f62056ee000 r--p 00028000 fd:01 56754381                   /usr/lib/x86_64-linux-gnu/libxcb.so.1.1.0
7f62056ee000-7f62056ef000 rw-p 00029000 fd:01 56754381                   /usr/lib/x86_64-linux-gnu/libxcb.so.1.1.0
…