@fabled 2025-03-17
Taking the idea further. Would it make sense to intern the strings globally? E.g. libc is mapped to practically all processes. So have a global `map` or LRU for the file names? Or is there some more generic way to intern a string in go?