@rockdaboot 2025-03-17
> Taking the idea further. Would it make sense to intern the strings globally? E.g. libc is mapped to practically all processes. So have a global `map` or LRU for the file names? Or is there some more generic way to intern a string in go?

Since Go 1.23 we have [the unique package](https://go.dev/blog/unique), but it comes with downsides (e.g. how to clean up?). An LRU Set makes more sense to me, as it cleans up automatically and the GC keeps track of strings that have been evicted but are still in use.