@rockdaboot 2025-03-17
> Taking the idea further. Would it make sense to intern the strings globally? E.g. libc is mapped to practically all processes. So have a global map or LRU for the file names? Or is there some more generic way to intern a string in go?

We can tackle this in a follow-up PR. Created an for a global de-deplication of strings to not forget about it: https://github.com/open-telemetry/opentelemetry-ebpf-profiler/issues/404